pyproject.toml
[project]
name = "walkie-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.0.0",
    "google-cloud-speech>=2.0.0",
    "httpx>=0.28.1",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "silero-vad>=5.0.0",
    "sounddevice>=0.4.6",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    "mcp>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
